I made the header which has the input field nested in it has a fixed position.// and the body tag has a top margin property with a value for the images to be fully visible.
added a header tag with a class of main-head// also i added a class to
the unordered list called pic-nav.//

in grey coupon there is a clear fix class that I added // added styles
to the input field ordered the pics how the mock up looks almost still
fixing that. made the lattice 100% in width.