the overlay has the img and content text show with a click of an image.
the overlay has a display: flex
and its center. also in the css the p tag has a colour of white which
is centered along with the image with a margin top.//

  html changed with the alt attr added to the img tag and deleted the
alt of the a tag//

 in the js file I changed the code which I added a var $para which
holds the content . and appended to the overlay before the image///
created a var called captionTex in the  function  that cache the string
that captures the alt attr of the img tag.